ENGR00161871: ENET: Changing MDIO read and write timeout
The original FEC_MII_TIMEOUT was set to 1ms, which is too low when
passed to the usecs_to_jiffies macro.

On ARM one jiffy is 10ms, so use msecs_to_jiffies instead and use
a timeout of 30ms, which corresponds to 3 jiffies.

After running 3 hours NFS boots, the MDIO timeout do not occur
anymore.

Signed-off-by: Rogerio Pimentel <rogerio.pimentel@freescale.com>