ENGR00283795: MX6: UART 1 hangs the system on SARBESD when not used a console
The platform hangs if UART 1 is not used a console on MX6 SABRESD.
The issue is because the UART UFCR register is not initialized and
the driver uses values provided by the bootloader  (U-Boot in this
case). This patch initializes the UFCR register and does not rely
on the default or what is configured by the bootloader.

Test method on MX6 SABRESD:
1. Add 'noconsole' to the kernel command line.
2. SSH to the board and type the below from the SSH window:
   echo Hello >/dev/ttymxc0
3. Platform will hang, with the patch the hang will not be seen.

Signed-off-by: Mahesh Mahadevan <Mahesh.Mahadevan@freescale.com>